Move folder-name-specific functionality to separate file
